Moves any non-component file to a new common package.
