+++ purchases/src/main/kotlin/com/revenuecat/purchases/paywalls/components/ImageComponent.kt
@@ -1,5 +1,7 @@
 package com.revenuecat.purchases.paywalls.components
 
+import com.revenuecat.purchases.paywalls.components.common.ComponentOverrides
+import com.revenuecat.purchases.paywalls.components.common.LocalizationKey
 import com.revenuecat.purchases.paywalls.components.properties.ColorScheme
 import com.revenuecat.purchases.paywalls.components.properties.FitMode
 import com.revenuecat.purchases.paywalls.components.properties.FitMode.FIT
