+++ purchases/src/main/kotlin/com/revenuecat/purchases/paywalls/components/PartialComponent.kt
@@ -1,0 +1,6 @@
+package com.revenuecat.purchases.paywalls.components
+
+/**
+ * Marker interface for partial components.
+ */
+internal sealed interface PartialComponent
