+++ purchases/src/main/kotlin/com/revenuecat/purchases/paywalls/components/TextComponent.kt
@@ -1,5 +1,7 @@
 package com.revenuecat.purchases.paywalls.components
 
+import com.revenuecat.purchases.paywalls.components.common.ComponentOverrides
+import com.revenuecat.purchases.paywalls.components.common.LocalizationKey
 import com.revenuecat.purchases.paywalls.components.properties.ColorScheme
 import com.revenuecat.purchases.paywalls.components.properties.FontSize
 import com.revenuecat.purchases.paywalls.components.properties.FontSize.BODY_M
