+++ purchases/src/main/kotlin/com/revenuecat/purchases/paywalls/components/common/Background.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.paywalls.components
+package com.revenuecat.purchases.paywalls.components.common
 
 import com.revenuecat.purchases.paywalls.components.properties.ColorScheme
 import com.revenuecat.purchases.paywalls.components.properties.ThemeImageUrls
