+++ purchases/src/test/java/com/revenuecat/purchases/paywalls/components/StickyFooterComponentTests.kt
@@ -1,6 +1,7 @@
 package com.revenuecat.purchases.paywalls.components
 
 import com.revenuecat.purchases.common.OfferingParser
+import com.revenuecat.purchases.paywalls.components.common.LocalizationKey
 import com.revenuecat.purchases.paywalls.components.properties.ColorInfo
 import com.revenuecat.purchases.paywalls.components.properties.ColorScheme
 import org.intellij.lang.annotations.Language
